ark dir

+++ TrainAndTest/Spectrogram/train_lores10_kaldi.py
@@ -129,7 +129,7 @@ parser.add_argument('--s', type=float, default=15, metavar='S',
                     help='the margin value for the angualr softmax loss function (default: 3.0')
 
 # args for a-softmax
-parser.add_argument('--m', type=float, default=3, metavar='M',
+parser.add_argument('--m', type=int, default=3, metavar='M',
                     help='the margin value for the angualr softmax loss function (default: 3.0')
 parser.add_argument('--lambda-min', type=int, default=5, metavar='S',
                     help='random seed (default: 0)')
